@@ -1,11 +1,13 @@
 package dev.fullstackcode.tc.docker.it;
 
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.testcontainers.containers.DockerComposeContainer;
+import org.testcontainers.containers.output.Slf4jLogConsumer;
 import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
@@ -26,10 +28,11 @@ public class BaseIT {
 
     static {
         environment =
-                new DockerComposeContainer(new File("src/test/resources/docker-compose.yaml"))
+                new DockerComposeContainer<>(new File("src/test/resources/docker-compose.yaml"))
                         .withExposedService("postgres", POSTGRES_PORT, Wait.forListeningPort())
+                        .withLogConsumer("postgres", new Slf4jLogConsumer(LoggerFactory.getLogger("testcontainers.postgres")));
 //                        .waitingFor("postgres", Wait.forLogMessage("started", 1))
-                        .withLocalCompose(true);
+//                        .withLocalCompose(true);
 
         environment.start();
         Runtime.getRuntime().addShutdownHook(new Thread(() -> stopContainer()));
