@@ -1,12 +1,16 @@
 package com.testcontainerspringboot.hero.universum;
 
+import org.slf4j.LoggerFactory;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.containers.output.Slf4jLogConsumer;
+import org.testcontainers.utility.DockerImageName;
 
 
 public abstract class DatabaseBaseTest {
-    static final MySQLContainer DATABASE = new MySQLContainer("mysql:latest");
+    static final MySQLContainer DATABASE = new MySQLContainer<>(DockerImageName.parse("mysql:8.0.33"))
+            .withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger("testcontainers.mysql")));
 
     static {
         DATABASE.start();
