@@ -1,17 +1,21 @@
 package com.testcontainerspringboot.hero.universum;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Collection;
 import org.junit.jupiter.api.Test;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.context.TestPropertySource;
 import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.containers.output.Slf4jLogConsumer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
+
+import java.util.Collection;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @Testcontainers
@@ -19,7 +23,8 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 class HeroClassicJpaRepositoryIT {
 
 	@Container
-	private static MySQLContainer database = new MySQLContainer("mysql:latest");
+	private static MySQLContainer database = new MySQLContainer<>(DockerImageName.parse("mysql:8.0.33"))
+			.withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger("testcontainers.mysql")));
 
 	@Autowired
 	private HeroClassicJpaRepository repositoryUnderTest;
