@@ -1,9 +1,7 @@
 package com.testcontainerspringboot.hero.universum;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Collection;
 import org.junit.jupiter.api.Test;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.util.TestPropertyValues;
@@ -11,56 +9,63 @@ import org.springframework.context.ApplicationContextInitializer;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.containers.output.Slf4jLogConsumer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
+
+import java.util.Collection;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @Testcontainers
 @ContextConfiguration(initializers = {HeroClassicJpaRepositoryIT.Initializer.class})
 class HeroClassicJpaRepositoryIT {
 
-	@Container
-	private static MySQLContainer mySQLContainer = (MySQLContainer) new MySQLContainer("mysql:latest")
-		.withDatabaseName("prop")
-		.withUsername("postgres")
-		.withPassword("pass")
-		.withExposedPorts(5432);
+    @Container
+    private static MySQLContainer mySQLContainer = new MySQLContainer<>(DockerImageName.parse("mysql:8.0.33"))
+            .withDatabaseName("prop")
+            .withUsername("postgres")
+            .withPassword("pass")
+            .withExposedPorts(5432)
+            .withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger("testcontainers.mysql")));
 
-	@Autowired
-	private HeroClassicJpaRepository repositoryUnderTest;
+    @Autowired
+    private HeroClassicJpaRepository repositoryUnderTest;
 
-	@Test
-	void findAllHero() {
-		int numberHeros = repositoryUnderTest.allHeros().size();
+    @Test
+    void findAllHero() {
+        int numberHeros = repositoryUnderTest.allHeros().size();
 
-		repositoryUnderTest.addHero(new Hero("Batman", "Gotham City", ComicUniversum.DC_COMICS));
-		repositoryUnderTest.addHero(new Hero("Superman", "Metropolis", ComicUniversum.DC_COMICS));
+        repositoryUnderTest.addHero(new Hero("Batman", "Gotham City", ComicUniversum.DC_COMICS));
+        repositoryUnderTest.addHero(new Hero("Superman", "Metropolis", ComicUniversum.DC_COMICS));
 
-		Collection<Hero> heros = repositoryUnderTest.allHeros();
+        Collection<Hero> heros = repositoryUnderTest.allHeros();
 
-		assertThat(heros).hasSize(numberHeros + 2);
-	}
+        assertThat(heros).hasSize(numberHeros + 2);
+    }
 
-	@Test
-	void findHeroByCriteria() {
-		repositoryUnderTest.addHero(new Hero("Batman", "Gotham City", ComicUniversum.DC_COMICS));
+    @Test
+    void findHeroByCriteria() {
+        repositoryUnderTest.addHero(new Hero("Batman", "Gotham City", ComicUniversum.DC_COMICS));
 
-		Collection<Hero> heros = repositoryUnderTest.findHerosBySearchCriteria("Batman");
+        Collection<Hero> heros = repositoryUnderTest.findHerosBySearchCriteria("Batman");
 
-		assertThat(heros).contains(new Hero("Batman", "Gotham City", ComicUniversum.DC_COMICS));
-	}
+        assertThat(heros).contains(new Hero("Batman", "Gotham City", ComicUniversum.DC_COMICS));
+    }
 
-	static class Initializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+    static class Initializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
 
-		@Override
-		public void initialize(ConfigurableApplicationContext applicationContext) {
-			TestPropertyValues.of(
-				"spring.datasource.url=" + mySQLContainer.getJdbcUrl(),
-				"spring.datasource.username=" + mySQLContainer.getUsername(),
-				"spring.datasource.password=" + mySQLContainer.getPassword(),
-				"spring.datasource.driverClassName=com.mysql.cj.jdbc.Driver",
-				"spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect")
-							  .applyTo(applicationContext);
-		}
-	}
+        @Override
+        public void initialize(ConfigurableApplicationContext applicationContext) {
+            TestPropertyValues.of(
+                            "spring.datasource.url=" + mySQLContainer.getJdbcUrl(),
+                            "spring.datasource.username=" + mySQLContainer.getUsername(),
+                            "spring.datasource.password=" + mySQLContainer.getPassword(),
+                            "spring.datasource.driverClassName=com.mysql.cj.jdbc.Driver",
+                            "spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect")
+                    .applyTo(applicationContext);
+        }
+    }
 }
